Convert html-script-finder_test to typescript
